Add in edit user + various fixes and tweaks
Adds a edit user function to UserHandler. Rather lengthy, so it's a
process that should be taken with care. Other changes:
- Added more documentation to my functions and a few in CryptoUtil
- Added a constant class to manage unavoidable xml constants
- Extended the testing code for the database handler
- Tweaked encrypt and decrypt: Instead of fully replacing the
functions for their global counterpart in UserHandler, a new
constructor was added. Create with only the file path, and the
class will default to using the universal one. The old constructor
will be deprecated soon, but I still need it to run the test code