Add test cases and adds stability to a few functions
Test cases for the handlers are now available in the test package.
More tests need to be done for edits, and more documentation needs
to be added to further explain some odd quirks of each function.
Other changes:
- UserHandler's checkUser can now accept null arguments, which check if
the file has no users. Read the comments for further explanation.
- .gitignore now ignore .xyt files, we don't need them in the
repository.